Crate imxrt_dma

Direct Memory Access (DMA) driver for i.MX RT processors

imxrt-dma provides

an unsafe API for defining and scheduling transfers with DMA Channels and Transfers

safe DMA futures for memcpy, peripheral-to-memory, and memory-to-peripheral transfers

This DMA driver may be re-exported from a HAL. If it is, you should consider using the safer APIs provided by your HAL.

Portability

This DMA driver works across all considered i.MX RT variants (1010 and 1060 family). You must make sure that the DMA channel you're creating is valid for your i.MX RT processor. This only matters on i.MX RT 1010 processors, which only support 16 DMA channels. Creating an invalid channel for your 1010 processor will result in a channel that references reserved memory.

License

Licensed under either of

Apache License, Version 2.0 (LICENSE-APACHE)

MIT License (LICENSE-MIT)

at your option.

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Modules

memcpy
DMA-powered memcpy

peripheral
DMA support for hardware peripherals

Structs

Channel
A DMA channel

Error
A wrapper around a DMA error status value

Transfer
The core DMA transfer future

Enums

BandwidthControl
Throttles the amount of bus bandwidth consumed by the eDMA

ChannelConfiguration
DMAMUX channel configuration

Traits

Element
Describes a transferrable DMA element; basically, an unsigned integer of any size.

Functions

on_error⚠
Handle a DMA error on one or more channels

on_interrupt⚠
Handle a DMA interrupt

set_destination_circular_buffer
Set a circular buffer as the destination for a DMA transfer

set_destination_hardware
Set a hardware peripheral as the destination for a DMA transfer

set_destination_linear_buffer
Set a linear buffer as the destination for a DMA transfer

set_source_circular_buffer
Set a circular buffer as the source for a DMA transfer

set_source_hardware
Set a hardware peripheral as the source for a DMA transfer

set_source_linear_buffer
Set a linear buffer as the source for a DMA transfer

Type Definitions

Result
A DMA result