Function imxrt_dma::channel::set_destination_linear_buffer

pub unsafe fn set_destination_linear_buffer<E: Element>(
    chan: &mut Channel,
    destination: &mut [E]
)

Set a linear buffer as the destination for a DMA transfer

When the transfer completes, the DMA channel will point at the start of the buffer.

Safety

Caller must ensure that the destination is valid for the lifetime of the transfer, and valid for all subsequent transfers performed by this DMA channel with this buffer.