Function imxrt_dma::on_error

pub unsafe fn on_error(max_channel: usize)

Handle a DMA error on one or more channels

on_error will find all DMA channels below max_channel that have an error. on_error then wakes the waker that channel's waker, if it exists.

max_channel is the total number of channels, starting from channel 0, that you want to check for errors. This should be 32, or 16 on a system with half the normal DMA channels. However, you may specify fewer channels if you only use a handful of channels.

Consider calling on_error in the DMA error interrupt:

use imxrt_dma::on_error;

// #[cortex_m_rt::interrupt]
fn DMA_ERROR() {
    // Safety: on an i.MX RT 1062 chip, there are
    // 32 DMA channels. This usage is valid.
    // This usage is NOT valid for an i.MX RT 1011
    // chip, which only has 16 DMA channels.
    unsafe { on_error(32) };
}

Safety

max_channel must be valid for your system. Specifying a value greater than 16 on a system that only has 16 channels results in undefined behavior.

Panics

Panics if max_channel is greater than 32.