Function imxrt_dma::on_interrupt

pub unsafe fn on_interrupt(channel: usize)

Handle a DMA interrupt

Checks the interrupt status for the channel identified by channel. If the channel completed its transfer, or it’s in an error state, on_interrupt wakes the channel’s waker.

Consider calling on_interrupt in a DMA channel’s interrupt handler:

use imxrt_dma::on_interrupt;

// #[cortex_m_rt::interrupt]
fn DMA7_DMA23() {
    // Safety: only checking channels 7 and 23, which
    // are both valid on an i.MX RT 1060 chip.
    unsafe {
        on_interrupt(7);
        on_interrupt(23);
    }
}

Safety

Caller must ensure that on_interrupt is called in the correct interrupt handler. Caller must ensure that channel is valid for the given system, and for the interrupt handler.

Panics

Panics if channel is greater than or equal to 32.