Function imxrt_dma::set_source_circular_buffer

pub fn set_source_circular_buffer<E: Element>(chan: &mut Channel, source: &[E])

Set a circular buffer as the source for a DMA transfer

When the transfer completes, the DMA channel remain at the next element in the circular buffer.

Panics

Panics if

the capacity is not a power of two

the alignment is not a multiple of the buffer's size in bytes