Function imxrt_dma::memcpy::memcpy

pub fn memcpy<'a, E: Element>(
    source: &'a [E], 
    destination: &'a mut [E], 
    channel: &'a mut Channel
) -> Memcpy<'a>

Perform a DMA-powered memcpy between the source and destination buffers

Copies the minimum number of elements between the two buffers. You're responsible for enabling any interrupts, and calling on_interrupt if the interrupt fires. Otherwise, you may poll the transfer until completion.

Example

Transfer 5 u32s between a source and destination buffer. The transfer completes when the DMA channel 7 interrupt fires.

use imxrt_dma::{Channel, memcpy, on_interrupt};

// #[cortex_m_rt::interrupt]
fn DMA7() {
    // Safety: DMA channel 7 valid
    unsafe { on_interrupt(7) };
}

let mut channel_7: Channel = // DMA channel 7
channel_7.set_interrupt_on_completion(true);
// TODO unmask DMA7 interrupt!

let source = [4u32, 5, 6, 7, 8];
let mut destination = [0; 5];

let transfer = memcpy::memcpy(&source, &mut destination, &mut channel_7);
executor::wfi(transfer);