imxrt_dma
Module channel

DMA channels

channel contains the DMA Channel type, along with helper functions for defining transfers.

Channel methods that specify memory involved in transfers are marked unsafe. You must be very careful when calling these methods, particuarly when a channel is already enabled.

Structs

Channel

A DMA channel

Enums

Configuration

DMAMUX channel configuration

Functions

set_destination_circular_buffer⚠

Set a circular buffer as the destination for a DMA transfer

set_destination_hardware⚠

Set a hardware peripheral as the destination for a DMA transfer

set_destination_linear_buffer⚠

Set a linear buffer as the destination for a DMA transfer

set_source_circular_buffer⚠

Set a circular buffer as the source for a DMA transfer

set_source_hardware⚠

Set a hardware peripheral as the source for a DMA transfer

set_source_linear_buffer⚠

Set a linear buffer as the source for a DMA transfer