Module imxrt_dma::peripheral

DMA support for hardware peripherals.

If a driver is compatible with this API, it implements some or all of the traits in this module. Consult your HAL for more information.

Each future documents when it resolves. To wake the executor, you can route the DMA channel’s interrupt handler to on_interrupt(). Otherwise, you can poll the future in a loop.

Structs

FullDuplex

A full-duplex DMA transfer from a single buffer

Read

A DMA transfer that receives data from hardware

Write

A DMA transfer that sends data to hardware

Traits

Bidirectional

Indicates that a peripheral can read and write from a single buffer using two simultaneous DMA transfers

Destination

A peripheral that can be the destination for DMA data

Source

A peripheral that can be the source of DMA data

Functions

full_duplex

Perform a full-duplex DMA transfer using two DMA channels that read and write from a single buffer.

read

Use a DMA channel to receive a buffer of elements from the source peripheral.

write

Use a DMA channel to send a buffer of data to the destination peripheral.