imxrt_dma::peripheral
Trait Bidirectional

pub unsafe trait Bidirectional<E: Element>: Source<E> + Destination<E> { }

Indicates that a peripheral can read and write from a single buffer using two simultaneous DMA transfers

It’s expected that the TX operation will drive the RX operation. A SPI transfer can be modeled in this manner.

Safety

Bidirectional assumes the same safety requirements as source and destination. Addtionally, you ensure that the peripheral is capable of this kind of transfer from a single buffer.