imxrt_dma::peripheral
Trait Destination

pub unsafe trait Destination<E: Element> {
    // Required methods
    fn destination_signal(&self) -> u32;
    fn destination_address(&self) -> *const E;
    fn enable_destination(&mut self);
    fn disable_destination(&mut self);
}

A peripheral that can be the destination for DMA data

By ‘destination,’ we mean that it receives data from a DMA transfer. A destination would be a peripheral that could send data out of processor memory, like a UART transmitter.

Safety

Destination should only be implemented on peripherals that are DMA capable. This trait should be implemented by HAL authors who are exposing DMA capable peripherals.

Required Methods

fn destination_signal(&self) -> u32

Peripheral destination request signal

See Table 4-3 of the reference manual. A destination mave has a qualifier like ‘transfer’ in the name.

fn destination_address(&self) -> *const E

Returns a pointer to the register into which the DMA channel writes data

This is the register that software writes to when sending data to a device. The type of the pointer describes the type of reads the DMA channel performs when transferring data.

fn enable_destination(&mut self)

Perform any actions necessary to enable DMA transfers

Callers use this method to put the peripheral into a state where it can accept transfers from a DMA channel.

fn disable_destination(&mut self)

Perform any actions necessary to disable or cancel DMA transfers

This may include undoing the actions in enable_destination.